requirements.txt
streamlit
pandas
numpy
Pillow
streamlit-drawable-canvas
